+++ changelog.md
@@ -7,6 +7,7 @@
 - Add more checks to get/set value from/tp values
 - Use more tuple instead of list
 - Cleanup code
+- Collect subscribed entities in central
 
 # Version 2023.10.4 (2023-10-03)